conversion to namespace step 06
