+++ classes/local/ipe/layout_insearchform.php
@@ -15,19 +15,16 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Contains class mod_surveypro\ipe_layout_required
+ * Contains class mod_surveypro\local\ipe\layout_insearchform
  *
  * @package   mod_surveypro
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_surveypro;
+namespace mod_surveypro\local\ipe;
 
 defined('MOODLE_INTERNAL') || die();
-
-use mod_surveypro\utility_layout;
-
 
 /**
  * Class to prepare an item variable for display and in-place editing
@@ -36,7 +33,7 @@ use mod_surveypro\utility_layout;
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class ipe_layout_required extends \core\output\inplace_editable {
+class layout_insearchform extends \core\output\inplace_editable {
 
     /**
      * @var sortindex
@@ -49,14 +46,14 @@ class ipe_layout_required extends \core\output\inplace_editable {
      * Constructor.
      *
      * @param int $itemid
-     * @param bool $required
+     * @param bool $insearchform
      * @param int $sortindex
      */
-    public function __construct($itemid, $required, $sortindex) {
+    public function __construct($itemid, $insearchform, $sortindex) {
         $this->sortindex = $sortindex;
 
-        $required = clean_param($required, PARAM_INT);
-        parent::__construct('mod_surveypro', 'layout_required', $itemid, true, '', $required);
+        $insearchform = clean_param($insearchform, PARAM_INT);
+        parent::__construct('mod_surveypro', 'layout_insearchform', $itemid, true, '', $insearchform);
         $this->set_type_toggle();
     }
 
@@ -68,15 +65,15 @@ class ipe_layout_required extends \core\output\inplace_editable {
      */
     public function export_for_template(\renderer_base $output) {
         if ($this->value) {
-            $requiredstr = get_string('requireditem_title', 'mod_surveypro');
-            $iconparams = array('id' => 'makeoptional_item_'.$this->sortindex);
-            $this->edithint = $requiredstr;
-            $this->displayvalue = $output->pix_icon('red', $requiredstr, 'mod_surveypro', $iconparams);
+            $insearchformstr = get_string('insearchform_title', 'mod_surveypro');
+            $iconparams = array('id' => 'removefromsearch_item_'.$this->sortindex);
+            $this->edithint = $insearchformstr;
+            $this->displayvalue = $output->pix_icon('insearch', $insearchformstr, 'mod_surveypro', $iconparams);
         } else {
-            $optionalstr = get_string('optionalitem_title', 'mod_surveypro');
-            $iconparams = array('id' => 'makerequired_item_'.$this->sortindex);
-            $this->edithint = $optionalstr;
-            $this->displayvalue = $output->pix_icon('green', $optionalstr, 'mod_surveypro', $iconparams);
+            $notinsearchformstr = get_string('notinsearchform_title', 'mod_surveypro');
+            $iconparams = array('id' => 'addtosearch_item_'.$this->sortindex);
+            $this->edithint = $notinsearchformstr;
+            $this->displayvalue = $output->pix_icon('notinsearch', $notinsearchformstr, 'mod_surveypro', $iconparams);
         }
 
         return parent::export_for_template($output);
@@ -86,29 +83,22 @@ class ipe_layout_required extends \core\output\inplace_editable {
      * Updates usertemplate name and returns instance of this object
      *
      * @param int $itemid
-     * @param string $newrequired
+     * @param string $newinsearchform
      * @return static
      */
-    public static function update($itemid, $newrequired) {
+    public static function update($itemid, $newinsearchform) {
         global $DB;
 
         $fields = 'id, surveyproid, type, plugin, sortindex';
         $itemrecord = $DB->get_record('surveypro_item', array('id' => $itemid), $fields, MUST_EXIST);
-        $surveypro = $DB->get_record('surveypro', array('id' => $itemrecord->surveyproid), 'id, course', MUST_EXIST);
+        $surveypro = $DB->get_record('surveypro', array('id' => $itemrecord->surveyproid), '*', MUST_EXIST);
         $cm = get_coursemodule_from_instance('surveypro', $surveypro->id, $surveypro->course, false, MUST_EXIST);
         $context = \context_module::instance($cm->id);
         \external_api::validate_context($context);
 
-        $tablename = 'surveypro'.$itemrecord->type.'_'.$itemrecord->plugin;
-        $newrequired = clean_param($newrequired, PARAM_INT);
-        $DB->set_field($tablename, 'required', $newrequired, array('itemid' => $itemid));
+        $newinsearchform = clean_param($newinsearchform, PARAM_INT);
+        $DB->set_field('surveypro_item', 'insearchform', $newinsearchform, array('id' => $itemid));
 
-        if (!empty($newrequired)) {
-            // This item that WAS NOT mandatory IS NOW mandatory.
-            $utilitylayoutman = new utility_layout($cm, $surveypro);
-            $utilitylayoutman->optional_to_required_followup($itemid);
-        }
-
-        return new static($itemid, $newrequired, $itemrecord->sortindex);
+        return new static($itemid, $newinsearchform, $itemrecord->sortindex);
     }
 }
