+++ classes/local/ipe/layout_required.php
@@ -15,16 +15,19 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Contains class mod_surveypro\ipe_layout_reserved
+ * Contains class mod_surveypro\local\ipe\layout_required
  *
  * @package   mod_surveypro
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_surveypro;
+namespace mod_surveypro\local\ipe;
 
 defined('MOODLE_INTERNAL') || die();
+
+use mod_surveypro\utility_layout;
+
 
 /**
  * Class to prepare an item variable for display and in-place editing
@@ -33,7 +36,7 @@ defined('MOODLE_INTERNAL') || die();
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class ipe_layout_reserved extends \core\output\inplace_editable {
+class layout_required extends \core\output\inplace_editable {
 
     /**
      * @var sortindex
@@ -46,14 +49,14 @@ class ipe_layout_reserved extends \core\output\inplace_editable {
      * Constructor.
      *
      * @param int $itemid
-     * @param bool $reserved
+     * @param bool $required
      * @param int $sortindex
      */
-    public function __construct($itemid, $reserved, $sortindex) {
+    public function __construct($itemid, $required, $sortindex) {
         $this->sortindex = $sortindex;
 
-        $reserved = clean_param($reserved, PARAM_INT);
-        parent::__construct('mod_surveypro', 'layout_reserved', $itemid, true, '', $reserved);
+        $required = clean_param($required, PARAM_INT);
+        parent::__construct('mod_surveypro', 'layout_required', $itemid, true, '', $required);
         $this->set_type_toggle();
     }
 
@@ -65,15 +68,15 @@ class ipe_layout_reserved extends \core\output\inplace_editable {
      */
     public function export_for_template(\renderer_base $output) {
         if ($this->value) {
-            $reservedstr = get_string('reserved_title', 'mod_surveypro');
-            $iconparams = array('id' => 'makeavailable_item_'.$this->sortindex);
-            $this->edithint = $reservedstr;
-            $this->displayvalue = $output->pix_icon('reserved', $reservedstr, 'surveypro', $iconparams);
+            $requiredstr = get_string('requireditem_title', 'mod_surveypro');
+            $iconparams = array('id' => 'makeoptional_item_'.$this->sortindex);
+            $this->edithint = $requiredstr;
+            $this->displayvalue = $output->pix_icon('red', $requiredstr, 'mod_surveypro', $iconparams);
         } else {
-            $availablestr = get_string('available_title', 'mod_surveypro');
-            $iconparams = array('id' => 'makereserved_item_'.$this->sortindex);
-            $this->edithint = $availablestr;
-            $this->displayvalue = $output->pix_icon('free', $availablestr, 'surveypro', $iconparams);
+            $optionalstr = get_string('optionalitem_title', 'mod_surveypro');
+            $iconparams = array('id' => 'makerequired_item_'.$this->sortindex);
+            $this->edithint = $optionalstr;
+            $this->displayvalue = $output->pix_icon('green', $optionalstr, 'mod_surveypro', $iconparams);
         }
 
         return parent::export_for_template($output);
@@ -83,22 +86,29 @@ class ipe_layout_reserved extends \core\output\inplace_editable {
      * Updates usertemplate name and returns instance of this object
      *
      * @param int $itemid
-     * @param string $newreserved
+     * @param string $newrequired
      * @return static
      */
-    public static function update($itemid, $newreserved) {
+    public static function update($itemid, $newrequired) {
         global $DB;
 
         $fields = 'id, surveyproid, type, plugin, sortindex';
         $itemrecord = $DB->get_record('surveypro_item', array('id' => $itemid), $fields, MUST_EXIST);
-        $surveypro = $DB->get_record('surveypro', array('id' => $itemrecord->surveyproid), '*', MUST_EXIST);
+        $surveypro = $DB->get_record('surveypro', array('id' => $itemrecord->surveyproid), 'id, course', MUST_EXIST);
         $cm = get_coursemodule_from_instance('surveypro', $surveypro->id, $surveypro->course, false, MUST_EXIST);
         $context = \context_module::instance($cm->id);
         \external_api::validate_context($context);
 
-        $newreserved = clean_param($newreserved, PARAM_INT);
-        $DB->set_field('surveypro_item', 'reserved', $newreserved, array('id' => $itemid));
+        $tablename = 'surveypro'.$itemrecord->type.'_'.$itemrecord->plugin;
+        $newrequired = clean_param($newrequired, PARAM_INT);
+        $DB->set_field($tablename, 'required', $newrequired, array('itemid' => $itemid));
 
-        return new static($itemid, $newreserved, $itemrecord->sortindex);
+        if (!empty($newrequired)) {
+            // This item that WAS NOT mandatory IS NOW mandatory.
+            $utilitylayoutman = new utility_layout($cm, $surveypro);
+            $utilitylayoutman->optional_to_required_followup($itemid);
+        }
+
+        return new static($itemid, $newrequired, $itemrecord->sortindex);
     }
 }
