+++ classes/local/ipe/layout_variable.php
@@ -15,14 +15,14 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Contains class mod_surveypro\ipe_layout_variable
+ * Contains class mod_surveypro\local\ipe\layout_variable
  *
  * @package   mod_surveypro
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_surveypro;
+namespace mod_surveypro\local\ipe;
 
 defined('MOODLE_INTERNAL') || die();
 
@@ -33,7 +33,7 @@ defined('MOODLE_INTERNAL') || die();
  * @copyright 2013 onwards kordan <kordan@mclink.it>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class ipe_layout_variable extends \core\output\inplace_editable {
+class layout_variable extends \core\output\inplace_editable {
     /**
      * Constructor.
      *
