+++ mtemplate_apply.php
@@ -28,7 +28,7 @@ use mod_surveypro\mastertemplate;
 use mod_surveypro\tabs;
 
 require_once(dirname(dirname(dirname(__FILE__))).'/config.php');
-require_once($CFG->dirroot.'/mod/surveypro/form/mtemplates/apply_form.php');
+require_once($CFG->dirroot.'/mod/surveypro/form/mtemplates/applyform.php');
 
 $id = optional_param('id', 0, PARAM_INT); // Course_module id.
 $s = optional_param('s', 0, PARAM_INT);   // Surveypro instance id.
