+++ view_export.php
@@ -26,7 +26,7 @@ use mod_surveypro\tabs;
 use mod_surveypro\view_export;
 
 require_once(dirname(dirname(dirname(__FILE__))).'/config.php');
-require_once($CFG->dirroot.'/mod/surveypro/form/data/export_form.php');
+require_once($CFG->dirroot.'/mod/surveypro/form/data/exportform.php');
 
 $id = optional_param('id', 0, PARAM_INT); // Course_module id.
 $s = optional_param('s', 0, PARAM_INT);   // Surveypro instance id.
